Train Barrier and Barrier length implementation to TrainType and fix the train client, train server, Train, siding java scripts errors in this change.

+++ src/main/java/mtr/data/Siding.java
@@ -240,7 +240,7 @@ public class Siding extends SavedRailBase implements IPacket {
 				trainsToSync.add(train);
 			}
 
-			if (train.closeToDepot(baseTrainType.getSpacing() * trainCars)) {
+			if (train.closeToDepot((int) baseTrainType.getSpacing() * trainCars)) {
 				spawnTrain = false;
 			}
